+++ src/DUPSS.API/Models/Common/CreateUserRequest.cs
@@ -1,0 +1,10 @@
+﻿using DUPSS.API.Models.Objects;
+
+namespace DUPSS.API.Models.Common
+{
+    public class CreateUserRequest
+    {
+        public required User User { get; set; }
+        public required string Password { get; set; }
+    }
+}
